brings us up to 22642cc0c360af9b10c53592249927f04ebb6832 (for the most part)

still needs playback fixed, adapter polling rate, and good config defaults

+++ Source/UnitTests/Common/FloatUtilsTest.cpp
@@ -1,6 +1,5 @@
 // Copyright 2018 Dolphin Emulator Project
-// Licensed under GPLv2+
-// Refer to the license.txt file included.
+// SPDX-License-Identifier: GPL-2.0-or-later
 
 #include <limits>
 #include <random>
@@ -29,10 +28,8 @@ TEST(FloatUtils, FlushToZero)
   // we want the multiplication to occur at test runtime.
   volatile float s = std::numeric_limits<float>::denorm_min();
   volatile double d = std::numeric_limits<double>::denorm_min();
-  // Casting away the volatile attribute is required in order for msvc to resolve this to the
-  // correct instance of the comparison function.
-  EXPECT_LT(0.f, (float)(s * 2));
-  EXPECT_LT(0.0, (double)(d * 2));
+  EXPECT_LT(0.f, s * 2);
+  EXPECT_LT(0.0, d * 2);
 
   EXPECT_EQ(+0.0, Common::FlushToZero(+std::numeric_limits<double>::denorm_min()));
   EXPECT_EQ(-0.0, Common::FlushToZero(-std::numeric_limits<double>::denorm_min()));
